Use individual bounding boxes for zoom extents.

+++ common/lc_math.h
@@ -1562,7 +1562,7 @@ inline void lcGetFrustumPlanes(const lcMatrix44& WorldView, const lcMatrix44& Pr
 	}
 }
 
-inline std::tuple<lcVector3, float> lcZoomExtents(const lcVector3& Position, const lcMatrix44& WorldView, const lcMatrix44& Projection, const lcVector3* Points, int NumPoints)
+inline std::tuple<lcVector3, float> lcZoomExtents(const lcVector3& Position, const lcMatrix44& WorldView, const lcMatrix44& Projection, const lcVector3* Points, size_t NumPoints)
 {
 	if (!NumPoints)
 		return std::make_tuple(Position, 2500.0f);
@@ -1580,7 +1580,7 @@ inline std::tuple<lcVector3, float> lcZoomExtents(const lcVector3& Position, con
 		const float ep = lcDot(Position, Plane);
 		const float fp = lcDot(Front, Plane);
 
-		for (int PointIdx = 0; PointIdx < NumPoints; PointIdx++)
+		for (size_t PointIdx = 0; PointIdx < NumPoints; PointIdx++)
 		{
 			const float u = (ep - lcDot(Points[PointIdx], Plane)) / fp;
 
@@ -1593,7 +1593,7 @@ inline std::tuple<lcVector3, float> lcZoomExtents(const lcVector3& Position, con
 
 	float FarDistance = 2500.0f;
 
-	for (int PointIdx = 0; PointIdx < NumPoints; PointIdx++)
+	for (size_t PointIdx = 0; PointIdx < NumPoints; PointIdx++)
 	{
 		const float Distance = lcDot(Points[PointIdx], Front);
